fix isBranch
JAL, JALR is not branch, but call. This bug caused infinite loops in
RISC-V branch selector. Since JAL, JALR was considered as a branch, the
branch selector tried to handle those instructions, but the changes were
made at the terminator instruction which is the end of the basic
block. This led to adding branch instruction again and again at the end
of the basic block