Fixed VAArg lowering in RV64
In RISC-V, 32-bit and 64-bit types are allowed in the instruction
selection stage. Var args are passed to the callee either
in register or in memory. Those passed in registers are also pushed onto
the stack in the callee. va_start() will set the pointer to the first
argument and each call of va_arg() will advance the pointer.
In RISCV64, both 32-bit and 64-bit types are pushed on the stack 8-byte
aligned. So the pointer must be advanced by 8 every time.

Signed-off-by: Jangseop Shin <jsshin@sor.snu.ac.kr>