Generate explicit sign and zero extension
RISCV-64 needs explicit sign extension for instructions without "W" at
the end, and explicit zero extension for MUL instruction. Before this
fix, "truncate" and "zext" IR instructions are selected to
EXTRACT_SUBREG and SUBREG_TO_REG, respectively, which were optimized
away during the following passes. This fix is not pretty, but I couldn't
think of a better way that's also simple to apply.

Signed-off-by: Jangseop Shin <jsshin@sor.snu.ac.kr>