Override isFPCloseToIncomingSP()
In current RISCV-LLVM, FP is actually set to the post-prolog SP,
pointing the bottom of the function frame. Emergency spill slot needs to
be close to the bottom of the stack, so that it can be reachble from FP
without using another register (that is, the offset is fit into the imm
field of LD/ST). If the offset does not fit, then we need a register to
load the offset into, so we need to find another emergency spill slot..
creating the infinite loop.
isFPCloseToIncomingSP() provides easy way to make the emergency spill
slot close to the top/bottom of the stack frame.

Signed-off-by: Jangseop Shin <jsshin@sor.snu.ac.kr>